requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
credentials==1.1
html5lib==1.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
lxml==4.6.1
mccabe==0.6.1
numpy==1.19.2
pandas==1.1.3
pkg-resources==0.0.0
pylint==2.6.0
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.11
urlopen==1.0.0
webencodings==0.5.1
wrapt==1.12.1
